New Convert class (task #7090)
* Created new \Qobo\Utils\Utility\Convert class
* moved valueToBytes() method from Utility to Convert class
* moved objectToArray() method from Utility to Convert class
* moved arrayToObject() method from Utility to Convert class
* Updated codebase (src/ and config/)
* Updated unit tests
