+++ src/ModuleConfig/Parser/V2/Json/ListParser.php
@@ -11,7 +11,7 @@
  */
 namespace Qobo\Utils\ModuleConfig\Parser\V2\Json;
 
-use Qobo\Utils\Utility;
+use Qobo\Utils\Utility\Convert;
 
 /**
  * List JSON Parser
@@ -52,7 +52,7 @@ class ListParser extends AbstractJsonParser
         $options = array_merge($this->options, $options);
 
         $result = parent::parse($path, $options);
-        $data = Utility::objectToArray($result);
+        $data = Convert::objectToArray($result);
         if (empty($data['items'])) {
             $data['items'] = [];
         }
@@ -66,7 +66,7 @@ class ListParser extends AbstractJsonParser
             $data['items'] = $this->flatten($data['items']);
         }
 
-        $result = Utility::arrayToObject($data);
+        $result = Convert::arrayToObject($data);
 
         return $result;
     }
