+++ tests/TestCase/ModuleConfig/Parser/V1/Csv/MigrationParserTest.php
@@ -4,7 +4,7 @@ namespace Qobo\Utils\Test\TestCase\ModuleConfig\Parser\V1\Csv;
 use Cake\Core\Configure;
 use PHPUnit\Framework\TestCase;
 use Qobo\Utils\ModuleConfig\Parser\V1\Csv\MigrationParser;
-use Qobo\Utils\Utility;
+use Qobo\Utils\Utility\Convert;
 
 class MigrationParserTest extends TestCase
 {
@@ -26,7 +26,7 @@ class MigrationParserTest extends TestCase
 
         $this->assertTrue(is_object($result), "Parser returned a non-object");
 
-        $result = Utility::objectToArray($result);
+        $result = Convert::objectToArray($result);
 
         $this->assertFalse(empty($result), "Parser returned empty result");
         $this->assertTrue(array_key_exists('id', $result), "Parser missed 'id' field");
@@ -45,7 +45,7 @@ class MigrationParserTest extends TestCase
         $result = $this->parser->parse($file);
 
         $this->assertTrue(is_object($result), "Parser returned a non-object");
-        $result = Utility::objectToArray($result);
+        $result = Convert::objectToArray($result);
 
         $this->assertTrue(empty($result), "Parser returned empty result");
     }
@@ -57,7 +57,7 @@ class MigrationParserTest extends TestCase
 
         $this->assertTrue(is_object($result), "Parser returned a non-object");
 
-        $result = Utility::objectToArray($result);
+        $result = Convert::objectToArray($result);
 
         $this->assertTrue(empty($result), "Parser returned non-empty result");
     }
@@ -68,7 +68,7 @@ class MigrationParserTest extends TestCase
         $result = $this->parser->parse($file);
 
         $this->assertTrue(is_object($result), "Parser returned a non-object");
-        $result = Utility::objectToArray($result);
+        $result = Convert::objectToArray($result);
 
         $this->assertFalse(empty($result), "Parser returned empty result");
     }
