+++ tests/TestCase/ModuleConfig/Parser/V1/Ini/ConfigParserTest.php
@@ -4,7 +4,7 @@ namespace Qobo\Utils\Test\TestCase\ModuleConfig\Parser\V1\Ini;
 use Cake\Core\Configure;
 use PHPUnit\Framework\TestCase;
 use Qobo\Utils\ModuleConfig\Parser\V1\Ini\ConfigParser;
-use Qobo\Utils\Utility;
+use Qobo\Utils\Utility\Convert;
 
 class ConfigParserTest extends TestCase
 {
@@ -26,7 +26,7 @@ class ConfigParserTest extends TestCase
 
         $this->assertTrue(is_object($result), "Parser returned a non-object");
 
-        $result = Utility::objectToArray($result);
+        $result = Convert::objectToArray($result);
 
         $this->assertFalse(empty($result), "Parser returned empty result");
 
@@ -52,7 +52,7 @@ class ConfigParserTest extends TestCase
 
         $this->assertTrue(is_object($result), "Parser returned a non-object");
 
-        $result = Utility::objectToArray($result);
+        $result = Convert::objectToArray($result);
 
         $this->assertFalse(empty($result), "Parser returned empty result");
 
@@ -68,7 +68,7 @@ class ConfigParserTest extends TestCase
 
         $this->assertTrue(is_object($result), "Parser returned a non-object");
 
-        $result = Utility::objectToArray($result);
+        $result = Convert::objectToArray($result);
 
         $this->assertArrayHasKey('associations', $result, "No associations found in the table config");
         $this->assertArrayHasKey('association_labels', $result['associations'], "No associations found in the table config");
@@ -90,7 +90,7 @@ class ConfigParserTest extends TestCase
         $this->assertTrue(is_array($warnings), "Warnings is not an array");
         $this->assertFalse(empty($warnings), "Warnings are empty");
 
-        $result = Utility::objectToArray($result);
+        $result = Convert::objectToArray($result);
 
         $this->assertFalse(empty($result['table']), "Parser missed 'table' section");
         $this->assertFalse(empty($result['table']['icon']), "Parser missed 'icon' default key");
