+++ tests/TestCase/ModuleConfig/Parser/V1/Ini/FieldsParserTest.php
@@ -5,7 +5,7 @@ use Cake\Core\Configure;
 use InvalidArgumentException;
 use PHPUnit\Framework\TestCase;
 use Qobo\Utils\ModuleConfig\Parser\V1\Ini\FieldsParser;
-use Qobo\Utils\Utility;
+use Qobo\Utils\Utility\Convert;
 
 class FieldsParserTest extends TestCase
 {
@@ -33,7 +33,7 @@ class FieldsParserTest extends TestCase
 
         $this->assertTrue(is_object($result), "Parser returned a non-object");
 
-        $result = Utility::objectToArray($result);
+        $result = Convert::objectToArray($result);
 
         $this->assertFalse(empty($result), "Parser returned empty result");
         $this->assertFalse(empty($result['cost']), "Parser missed 'cost' section");
