+++ tests/TestCase/ModuleConfig/Parser/V2/Json/ListParserTest.php
@@ -4,7 +4,7 @@ namespace Qobo\Utils\Test\TestCase\ModuleConfig\Parser\V2\Json;
 use Cake\Core\Configure;
 use PHPUnit\Framework\TestCase;
 use Qobo\Utils\ModuleConfig\Parser\V2\Json\ListParser;
-use Qobo\Utils\Utility;
+use Qobo\Utils\Utility\Convert;
 
 class ListParserTest extends TestCase
 {
@@ -31,7 +31,7 @@ class ListParserTest extends TestCase
         $file = $this->dataDir . DS . 'Foo' . DS . 'lists' . DS . 'local_genders.json';
         $result = $this->parser->parse($file);
 
-        $resultArray = Utility::objectToArray($result);
+        $resultArray = Convert::objectToArray($result);
 
         $this->assertNotEmpty($result);
         $this->assertArrayHasKey('items', $resultArray);
@@ -43,7 +43,7 @@ class ListParserTest extends TestCase
         $file = $this->dataDir . DS . 'Foo' . DS . 'lists' . DS . 'local_genders.json';
         $result = $this->parser->parse($file, ['filter' => true]);
 
-        $resultArray = Utility::objectToArray($result);
+        $resultArray = Convert::objectToArray($result);
         $this->assertTrue(!in_array('foo', array_keys($resultArray['items'])));
     }
 
@@ -52,7 +52,7 @@ class ListParserTest extends TestCase
         $file = $this->dataDir . DS . 'Foo' . DS . 'lists' . DS . 'local_genders.json';
         $result = $this->parser->parse($file, ['filter' => true, 'flatten' => true]);
 
-        $resultArray = Utility::objectToArray($result);
+        $resultArray = Convert::objectToArray($result);
 
         $this->assertTrue(in_array('bar.bar_one', array_keys($resultArray['items'])));
         $this->assertTrue(in_array('bar.bar_two', array_keys($resultArray['items'])));
